Change from Google Cloud to Firebase hosting.
With Google Cloud Storage there is no way to obtain a static IP
to a bucket so that you could create a DNS "A" record to it.
Or at least I did not find it after hours of reading through
their docs and examples. In Firebase hosting I got a DNS "A"
record to my application in about a minute.